Move packets out of data package
