+++ src/main/java/com/mraof/minestuck/network/TitleDataPacket.java
@@ -1,7 +1,6 @@
-package com.mraof.minestuck.network.data;
+package com.mraof.minestuck.network;
 
 import com.mraof.minestuck.Minestuck;
-import com.mraof.minestuck.network.MSPacket;
 import com.mraof.minestuck.player.ClientPlayerData;
 import com.mraof.minestuck.player.Title;
 import net.minecraft.network.FriendlyByteBuf;
